improv(ChangeDetection): Run render outside angular zone since it doesn't need to trigger change detection
As of now, rendering the chart was done in the angular zone. This meant that all events would be used by ngZone to trigger change detection. Therefore, things like the hover for tooltips, requestAnimationFrames for animations and others were triggering ChangeDetection. This causes a major impact on performance when having multiple charts