fix: make SSR compatible
In this commit, we have adjusted the chart to render only in the browser. Additionally, we need
to lazy load the ApexCharts library in the browser as well. This is necessary because the library
statically assumes the presence of the window object without checking whether it is available,
leading to an error stating that "window is not defined".